+++ labbook_BE/default_settings.py
@@ -1,6 +1,6 @@
 # -*- coding:utf-8 -*-
 
-APP_VERSION = '3.1.0'
+APP_VERSION = '3.1.1'
 
 # ######################################
 #  Database parameters